tools/pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Tools library for the Aden agent framework"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Aden", email = "team@aden.ai" }]
keywords = ["ai", "agents", "tools", "llm"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "pypdf>=4.0.0",
    "pandas>=2.0.0",
    "jsonpath-ng>=1.6.0",
    "fastmcp>=2.0.0",
    "diff-match-patch>=20230430",
    "python-dotenv>=1.0.0",
    "playwright>=1.40.0",
    "playwright-stealth>=1.0.5",
    "litellm>=1.81.0",
    "dnspython>=2.4.0",
    "resend>=2.0.0",
    "google-analytics-data>=0.18.0",
    "framework",
    "stripe>=14.3.0",
    "arxiv>=2.1.0",
    "requests>=2.31.0",
    "psycopg2-binary>=2.9.0",
]

[project.scripts]
aden-office-pack = "aden_tools.cli.office_pack:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pyright>=1.1.380",
    "pre-commit>=3.7.0",
]
sandbox = [
    "RestrictedPython>=7.0",
]
ocr = [
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
]
excel = [
    "openpyxl>=3.1.0",
]
powerpoint = [
    "python-pptx>=1.0.0",
]
word = [
    "python-docx>=1.1.0",
]
charts = [
    "matplotlib>=3.8.0",
]
sql = [
    "duckdb>=1.0.0",
]
bigquery = [
    "google-cloud-bigquery>=3.0.0",
]
all = [
    "RestrictedPython>=7.0",
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
    "duckdb>=1.0.0",
    "openpyxl>=3.1.0",
    "google-cloud-bigquery>=3.0.0",
    "python-pptx>=1.0.0",
    "matplotlib>=3.8.0",
    "pyright>=1.1.380",
    "pre-commit>=3.7.0",


]

[tool.uv.sources]
framework = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aden_tools"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["aden_tools"]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "ty>=0.0.13",
    "ruff>=0.14.14",
    "duckdb>=1.4.4",
]
powerpoint = [
    "python-pptx>=1.0.0",
]
word = [
    "python-docx>=1.1.0",
]
charts = [
    "matplotlib>=3.8.0",
]

